Apply bucket policy directly to bucket resource
Prior to this change, attempts to create a cloudtrail using this logs
bucket would fail initially with an InsufficientS3BucketPolicyException.
It was an ordering issue, where the cloudtrail would attempt to use the
s3 bucket before the subsequent policy attachment was complete. Running
a second terraform apply would work, but the logs bucket and cloudtrail
should be able to be created in the same terraform apply.

This change removes the aws_s3_bucket_policy resource and applies the
bucket policy directly in the aws_s3_bucket resource.

https://www.pivotaltracker.com/story/show/169575508