Add option of enforcing TLS-only requests via the S3 bucket policy
This change allows an extra statement to be added to the log S3 bucket
policy that denies requests unless they're made via TLS.

This additional policy statement is necessary to be able to enforce
encryption-in-transit requirements alongside the encryption-at-rest
that's already handled by the
`apply_server_side_encryption_by_default` block on the S3 bucket
resource.

This policy is required for buckets managed by this module to pass the
AWS Config `s3-bucket-ssl-requests-only`[1] managed rule and
Prowler's[2] `extra764` check, which although not part of the CIS
benchmark seems like a sensible thing to satisfy.

Note that Trussworks' AWS Config module[3] enables the
s3-bucket-ssl-requests-only check by default, so before this change,
the two modules together result in failing Config checks.

By default insecure requests are permitted, in line with S3 defaults
and the historical behaviour of this module.

Note that this change does not appear to be implementable in a safe
manner following the pattern of other statements of flipping the
effect. My reading of explicitly allowing aws:SecureTransport requests
would have the effect of permitting all anonymous HTTP requests, which
is definitely not the opposite of denying any HTTP requests.

[1] https://docs.aws.amazon.com/config/latest/developerguide/s3-bucket-ssl-requests-only.html
[2] https://github.com/toniblyx/prowler
[3] https://github.com/trussworks/terraform-aws-config