Enforce TLS-only requests to S3 via bucket policy by default
This change adds an extra statement to the log S3 bucket policy that
denies requests unless they're made via TLS.

This additional policy statement is necessary to be able to enforce
encryption-in-transit requirements alongside the encryption-at-rest
that's already handled by the
`apply_server_side_encryption_by_default` block on the S3 bucket
resource.

This policy is required for buckets managed by this module to pass the
AWS Config `s3-bucket-ssl-requests-only`[1] managed rule and
Prowler's[2] `extra764` check, which although not part of the CIS
benchmark seems like a sensible thing to satisfy.

Note that Trussworks' AWS Config module[3] enables the
s3-bucket-ssl-requests-only check by default, so before this change,
the two modules together result in failing Config checks.

A previous PR was submitted which made this additional policy
optional, with the default unchanged, but it was agreed with @dynamike
that changing the default was more sensible.

[1] https://docs.aws.amazon.com/config/latest/developerguide/s3-bucket-ssl-requests-only.html
[2] https://github.com/toniblyx/prowler
[3] https://github.com/trussworks/terraform-aws-config